/: Genesis
Add all the files required for the building and maintaining of the site
in its current initial state. At the moment the site is nothing but a
homepage with some (pretty nice) CSS, a fully functioning elisp build
script, and some other stuff.

The `syntax.css' file does contain styles for stuff not yet found on the
site, but thats because I already have another page under `doc/' in the
works.