build: Add support for exporting tables
ox-slimhtml does not have support for exporting tables, so we
implement it manually. The first row of the table is going to be the
heading row, so we can use a regular expression to substitue the <tr>
tags to <th> tags. It works because exported tables look like this:

  <table><tr><td>Some Text</td></tr>
  <tr><td>More Text</td></tr>
  ...

Like the previous commit, tables are not used yet, but they will be shortly.

TODO: Add styling for tables