build: Patch pup(1) bug inserting extra whitespace
When ending a sentence with formatted text, you would want the
generated HTML to look something like this:

  <p><em>This is emphatic</em>!</p>

This ensures that the word "emphatic" is followed directly by an
exclemation mark. With pup(1)'s pretty printing however, the following
HTML can be generated:

  <p>
      <em>
          This is emphatic
      </em>
      !
  </p>

This is bad, because when rendered on the screen a whitespace
character is placed in between the word "emphatic" and the exclamation
mark giving us "This is emphatic !" instead of the desired "This is
emphatic!".

Additionally, since we no longer use pup(1), the DOCTYPE is not
removed from the top of the document anymore, so the call to sed(1) is
no longer needed.