/etc/doc: Add a few PDFs I found online
This commit includes 4 PDFs covering the following topics:

 - Awk
 - BC
 - DC
 - Structural Regular Expressions

The first three are introductions to the Awk, BC, and DC lanauges which
I feel are not documented well enough, BC and DC especially. The fourth
PDF covers the idea of structural regular expressions and I personally
believe it's a very interesting read.